+++ lib/src/21_base32.dart
@@ -70,23 +70,30 @@ abstract class RandomBase32 implements Random {
         : combineUpper53bitsJS(nextRaw32(), nextRaw32());
   }
 
+  static final int _POW2_32 = 4294967296; // it's (1 << 32). For JS it's safer to set a constant
+
   /// Generates a non-negative random integer uniformly distributed in
   /// the range from 0, inclusive, to [max], exclusive.
-  ///
-  /// To make the distribution uniform, we use the so-called
-  /// [Debiased Modulo Once - Java Method](https://git.io/Jm0D7).
-  ///
-  /// This implementation is slightly faster than the standard one for
-  /// all [max] values, except for [max], which are powers of two.
   @override
   int nextInt(int max) {
-    if (max < 1 || max > 0xFFFFFFFF) {
-      throw RangeError.range(max, 1, 0xFFFFFFFF);
+    // almost the same as https://bit.ly/35OH1Vh
+
+    if (max <= 0 || max > _POW2_32) {
+      throw RangeError.range(
+          max, 1, _POW2_32, 'max', 'Must be positive and <= 2^32');
     }
-    int r = nextRaw32();
-    int m = max - 1;
-    for (int u = r; u - (r = u % max) + m < 0; u = nextRaw32()) {}
-    return r;
+    if ((max & -max) == max) {
+      // Fast case for powers of two.
+      return nextRaw32() & (max - 1);
+    }
+
+    int rnd32;
+    int result;
+    do {
+      rnd32 = nextRaw32();
+      result = rnd32 % max;
+    } while ((rnd32 - result + max) > _POW2_32);
+    return result;
   }
 
   /// Generates a random floating point value uniformly distributed
@@ -148,3 +155,4 @@ abstract class RandomBase32 implements Random {
   @protected
   int boolCache_prevShift = 0;
 }
+