+++ lib/xrandom.dart
@@ -11,4 +11,6 @@ export 'src/60_xorshift32.dart' show Xorshift32;
 export 'src/60_xorshift64.dart' show Xorshift64;
 export 'src/60_xoshiro128pp.dart' show Xoshiro128pp;
 export 'src/60_xoshiro256.dart' show Xoshiro256pp, Xoshiro256ss;
-export 'src/90_aliases.dart' show Xrandom, Qrandom, Drandom;
+
+export 'src/80_drandom.dart' show Drandom;
+export 'src/90_aliases.dart' show Xrandom, Qrandom;